Disable Inspection
Disabling the inspect tool and F12 key is generally not recommended as it can be easily bypassed by users and may lead to a poor user experience. However, if you still want to implement this for specific reasons, you can use JavaScript to block the right-click context menu and disable certain key combinations.